Extract logic to import transactions into YNAB
